Transcript_bot_baсkup/requirements.txt
python-telegram-bot==20.8
openai==1.14.1
python-dotenv==1.0.1
yt-dlp==2024.3.10
moviepy==1.0.3
pydub==0.25.1
markdown==3.5.2
tiktoken==0.6.0
aiofiles==23.2.1
whisper==1.1.10
requests==2.31.0
beautifulsoup4==4.12.3
